Relocate data for HQ & plan page plan section
Moved the data needed for the headquarters and plan page plan section
modules to their own data modules within their respective dirs. Makes
for cleaner components.
